requirements.txt
biopython==1.84
biotite==1.0.1
matplotlib==3.7.1
networkx==3.2.1
numpy==1.23.5
pandas==2.2.3
Requests==2.32.3
scikit_learn==1.6.0
scipy==1.14.1
torch==2.5.1+cu121
torch_geometric==2.6.1
